Rebase from 'upstream'

+++ yasmin_ros/src/yasmin_ros/get_parameters_state.cpp
@@ -1,17 +1,16 @@
 // Copyright (C) 2025 Miguel Ángel González Santamarta
 //
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 #include "yasmin_ros/get_parameters_state.hpp"
 
@@ -74,7 +73,7 @@ GetParametersState::execute(yasmin::Blackboard::SharedPtr blackboard) {
       } else if (default_value.type() == typeid(std::vector<bool>)) {
         this->node_->declare_parameter(
             param_name, std::any_cast<std::vector<bool>>(default_value));
-      } else if (default_value.type() == typeid(std::vector<int>)) {
+      } else if (default_value.type() == typeid(std::vector<int64_t>)) {
         this->node_->declare_parameter(
             param_name, std::any_cast<std::vector<int64_t>>(default_value));
       } else if (default_value.type() == typeid(std::vector<double>)) {
